mysql

+++ fuck_game_server_engine/const.h
@@ -2,5 +2,5 @@
 
 const uint32_t c_ip_size = 24;
 
-const float c_pi = 3.1415926535;
-const float c_min_float = 1.0E-9;
+const double c_pi = 3.1415926535;
+const double c_min_float = 1.0E-9;
